% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part2.R
\name{allele_dose}
\alias{allele_dose}
\title{Count HLA allele (allele dose) of interest and display their distribution across group}
\usage{
allele_dose(file, selection, outputFile)
}
\arguments{
\item{file}{The File containing Hla allele for each individuals and probability to have it}

\item{selection}{vector of allele of interest}

\item{outputFile}{The file that will contain the input info and}
}
\value{
file with with all input file's data and 2xselection colunms: one for allele count(0 1 2 ) and the other with fractionned dose
}
\description{
Take a file containing at least colunms: subjectId , allele1 , allele 2, prob
}
