% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part1.R
\name{firstQC}
\alias{firstQC}
\title{Do a standardised quality control, it cleans your data by removing not informative's SNP and analyze stratification of your population}
\usage{
firstQC(bedFile = NULL, bimFile = NULL, famFile = NULL, vcf = NULL,
  outputFolder, outputFile)
}
\arguments{
\item{bedFile}{.bed file containing the packed binary SNP genotype data}

\item{bimFile}{.bim file containingthe SNP descriptions}

\item{famFile}{.fam file containing subject (and, possibly, family) identifiers. This is basically a tab-delimited "pedfile"}

\item{vcf}{variant call format file}

\item{outputFolder}{folder where to save the result}

\item{outputFile}{Name wanted for your VCF}
}
\value{
Write a summary of the data after the quality control and cleaned file
}
\description{
Do a standardised quality control, it cleans your data by removing not informative's SNP and analyze stratification of your population
}
