% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part3.R
\name{hlaAlleleAnalysis}
\alias{hlaAlleleAnalysis}
\title{Analyze association of HLA alleles with many variate phenotypes (like case/control or quantitative trait)}
\usage{
hlaAlleleAnalysis(
  dataPath,
  phenoPath = NULL,
  covarPath = NULL,
  outputFolder,
  outputFile,
  type = "pheno"
)
}
\arguments{
\item{dataPath}{HLA alleles file}

\item{phenoPath}{phenotypes file with different trait's names in columns}

\item{covarPath}{must contains SubjectID followed by covariables}

\item{outputFolder}{folder where to save the result}

\item{outputFile}{result's filename}

\item{type}{either "disease" for case/control analysis or "pheno" with quantitative phenotype(s)}
}
\value{
Linear Regression result and many plots : QQplot, personnalized manhattan plot and/or variation of phenotype's expression
}
\description{
Take as input :
(1) a file containing the following 14 colunms :  subjectID Disease PC1 PC2  A A B B C C DRB1 DRB1 DQB1 DQB1 ("," separator between fields)
(2 optional) file with differents quantitative phenotypes : subjectID traitName1 traitName2 ... traitName(n)
(3 optional) file with covariable to correct regressions
}
