% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part3.R
\name{hlaCexpressionAnalysis}
\alias{hlaCexpressionAnalysis}
\title{HLA-C expression analys many variate phenotypes}
\usage{
hlaCexpressionAnalysis(
  dataPath,
  phenoPath = NULL,
  statut = NULL,
  pcaPath = NULL,
  covarPath = NULL,
  outputFolder,
  outputFile,
  type = "pheno"
)
}
\arguments{
\item{dataPath}{file with subject.ID and HLA-c expression "quantitative"}

\item{phenoPath}{file with subject.ID and as many columns as quantitative phenotypes}

\item{statut}{file conaining subject ID and Disease statut}

\item{pcaPath}{pca result for PC1 and PC2}

\item{covarPath}{must contains SubjectID followed by covariables}

\item{outputFolder}{folder where to save the result}

\item{outputFile}{Name for the result csv file with columns: subjectID    Disease    PC1    PC2    A    A    B    B    C    C    DRB1    DRB1    DQB1    DQB1}

\item{type}{etiher "disease" for analysis in case/control or "pheno" for analysis with an quantitative phenotype}
}
\value{
Linear Regression result with the phenotypes or case/control and corrected either with PCs or with covariates
}
\description{
HLA-C expression analys many variate phenotypes
}
