% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part3.R
\name{hlaHaplotypeAnalysis}
\alias{hlaHaplotypeAnalysis}
\title{Analysis of HLA haplotypes and many phenotypes}
\usage{
hlaHaplotypeAnalysis(haploDataPath, statut, phenoPath = NULL,
  pcaPath = NULL, covarPath = NULL, outputFolder, outputFile,
  type = "pheno")
}
\arguments{
\item{haploDataPath}{Hla haplotypes file pathways}

\item{statut}{file conaining subject ID and Disease statut}

\item{phenoPath}{phenotypes file with different trait names in columns}

\item{pcaPath}{pca result for PC1 and PC2}

\item{covarPath}{must contains SubjectID followed by covariables}

\item{outputFolder}{folder where to save the result}

\item{outputFile}{Name for the result csv file with columns: subjectID    Disease    PC1    PC2    A    A    B    B    C    C    DRB1    DRB1    DQB1    DQB1}

\item{type}{etiher "disease" for analysis in case/control or "pheno" for analysis with an quantitative phenotype}
}
\value{
Linear Regression result with the phenotypes as covariates
}
\description{
Analysis of HLA haplotypes and many phenotypes
}
