% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part2.R
\name{hlaImputation}
\alias{hlaImputation}
\title{Impute HLA allele by using HiBAG package}
\usage{
hlaImputation(bedFile, bimFile, famFile, hlaLoci = c("A", "B", "C",
  "DRB1", "DQB1"), trainingModel = NULL, pcaFile, outputFolder,
  probThreshold = 0.5)
}
\arguments{
\item{bedFile}{.bed file containing the packed binary SNP genotype data}

\item{bimFile}{.bim file containingthe SNP descriptions}

\item{famFile}{.fam file containing subject (and, possibly, family) identifiers. This is basically a tab-delimited "pedfile"}

\item{hlaLoci}{vector of HLA loci, by default it is: A B C DPR1}

\item{trainingModel}{Model for hibag training. Download pre-fit Model https://github.com/zhengxwen/HIBAG}

\item{pcaFile}{FIle with two principal components of PCA}

\item{outputFolder}{folder where to save the result}

\item{probThreshold}{Threshold of Postprob from Hibag to keep. By default 0.5}
}
\value{
All HLA alleles with imputation probabilities
}
\description{
Impute HLA allele by using HiBAG package
}
