% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part1.R
\name{plink2vcf}
\alias{plink2vcf}
\title{Check and fix strand, check reference alleles and remove duplicate. These QC are recommended by Sanger server imputation: wrayner@well.ox.ac.uk}
\usage{
plink2vcf(dataFolder, plinkFile, outputFile, outputFolder)
}
\arguments{
\item{dataFolder}{folder containing bed/bim/fam and .frq files}

\item{plinkFile}{bed/bim/fam file's name without extension and without pathways}

\item{outputFile}{Name wanted for your VCF}

\item{outputFolder}{folder where to save the resul}
}
\value{
VCF file ready for SNP imputation
}
\description{
Check and fix strand, check reference alleles and remove duplicate. These QC are recommended by Sanger server imputation: wrayner@well.ox.ac.uk
}
