% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part1.R
\name{popStrat}
\alias{popStrat}
\title{Checking for population stratification in order to adjust and get homogenous sample}
\usage{
popStrat(bedFile, bimFile, famFile, workingFolder)
}
\arguments{
\item{bedFile}{.bed file containing the packed binary SNP genotype data}

\item{bimFile}{.bim file containingthe SNP descriptions}

\item{famFile}{.fam file containing subject (and, possibly, family) identifiers. This is basically a tab-delimited "pedfile"}

\item{workingFolder}{folder where to save the result}
}
\value{
Return a plot sample's individuals ancestry with 3 references populations: European (CEU) , East-Asian (CHB) and African (YRI)
}
\description{
Checking for population stratification in order to adjust and get homogenous sample
}
