% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part3.R
\name{hlaAlleleAnalysis}
\alias{hlaAlleleAnalysis}
\title{Analysis of HLA allele and many variate phenotypes}
\usage{
hlaAlleleAnalysis(dataPath, phenoPath = NULL, covarPath = NULL,
  outputFolder, outputFile, type = "pheno")
}
\arguments{
\item{dataPath}{Hla allele file pathways. Contains  with columns: subjectID    Disease    PC1    PC2    A    A    B    B    C    C    DRB1    DRB1    DQB1    DQB1.}

\item{phenoPath}{phenotypes file with different trait names in columns}

\item{covarPath}{must contains SubjectID followed by covariables}

\item{outputFolder}{folder where to save the result}

\item{outputFile}{Name for the result csv file}

\item{type}{etiher "disease" for analysis in case/control or "pheno" for analysis with an quantitative phenotype}
}
\value{
Linear Regression result with the phenotypes as covariates
}
\description{
Take a file containing 14 colunms as input
     subjectID Disease PC1 PC2  A A B B C C DRB1 DRB1 DQB1 DQB1 ("," separator between fields)
AND a file with differents phenotypes :
     subjectID traitName1 traitName2 traitName3 traitName4 traitName5
}
