% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part3.R
\name{hlaKIRAnalysis}
\alias{hlaKIRAnalysis}
\title{HLA KIR ligand analysis with multiple phenotypes or in case/control}
\usage{
hlaKIRAnalysis(datapath, phenoPath = NULL, statut, pcaPath = NULL,
  covarPath = NULL, outputFolder, outputFile, type = "pheno")
}
\arguments{
\item{datapath}{File from Easy-HLA}

\item{phenoPath}{phenotypes file with different trait names in columns}

\item{statut}{file conaining subject ID and Disease statut}

\item{pcaPath}{pca result for PC1 and PC2}

\item{covarPath}{must contains SubjectID followed by covariables}

\item{outputFolder}{Folder's pathways where to save the result file}

\item{outputFile}{Name of the result file}

\item{type}{etiher "disease" for analysis in case/control or "pheno" for analysis with an quantitative phenotype}
}
\value{
Regression result in a file with columns: Amino_Acid	Frequency	Frequency_control	Frequency_case	N	P_allelic	OR_allelic	lower_95%_CI_allelic	upper_95%_CI_allelic	Beta_allelic	SE_allelic	P_dominant	OR_dominant	lower_95%_CI_dominant	upper_95%_CI_dominant	Beta_dominant	SE_dominant
}
\description{
HLA KIR ligand analysis with multiple phenotypes or in case/control
}
